% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_processing.R
\name{calculate_best_response}
\alias{calculate_best_response}
\title{Calculate Best Overall Response}
\usage{
calculate_best_response(data)
}
\arguments{
\item{data}{Processed clinical data}
}
\value{
Data frame with best overall response per patient
}
\description{
Calculate Best Overall Response
}
