% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_processing.R
\name{generate_data_summary}
\alias{generate_data_summary}
\title{Generate Data Summary Statistics}
\usage{
generate_data_summary(data)
}
\arguments{
\item{data}{Processed clinical data}
}
\value{
List with summary statistics
}
\description{
Generate Data Summary Statistics
}
