% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_processing.R
\name{process_structured_data}
\alias{process_structured_data}
\title{Process Already Structured Clinical Data}
\usage{
process_structured_data(
  data,
  patient_col,
  treatment_start_col,
  assessment_date_col,
  response_col,
  treatment_end_col = NULL,
  treatment_type_col = NULL
)
}
\arguments{
\item{data}{Structured clinical data frame}

\item{patient_col}{Name of patient ID column}

\item{treatment_start_col}{Name of treatment start column}

\item{assessment_date_col}{Name of assessment date column}

\item{response_col}{Name of response column}

\item{treatment_end_col}{Name of treatment end column (optional)}

\item{treatment_type_col}{Name of treatment type column (optional)}
}
\value{
Standardized data frame
}
\description{
Process Already Structured Clinical Data
}
